requirements.txt
aiogram
python-dotenv
openrouter
